renames container and component files
